tg_bot/requirements.txt
anyio==4.1.0
APScheduler==3.10.4
black==23.11.0
certifi==2023.11.17
click==8.1.7
colorama==0.4.6
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.11.2
platformdirs==4.1.0
python-dotenv==1.0.0
python-telegram-bot==20.7
requests==2.31.0
sniffio==1.3.0
